+++ src/app/_hooks/useFunnel/_component/Step.tsx
@@ -1,0 +1,9 @@
+import { ReactElement } from "react";
+
+import { StepProps } from "../types";
+
+const Step = ({ children }: StepProps): ReactElement => {
+  return <>{children}</>;
+};
+
+export default Step;
