+++ src/app/_hooks/useFunnel/types.ts
@@ -1,0 +1,11 @@
+import { ReactElement, ReactNode } from "react";
+
+export interface FunnelProps {
+  step: string;
+  children: Array<ReactElement<StepProps>>;
+}
+
+export interface StepProps {
+  name: string;
+  children: ReactNode;
+}
